move rspc frontend code into a `utils/rspc.ts` file
