+++ create-rspc-app/templates/react_base_axum/src/main.tsx
@@ -1,12 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import rspc, { client, queryClient } from "./query.axum";
+import { rspc, client, queryClient } from "./utils/rspc";
 
-ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
+ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <rspc.Provider client={client} queryClient={queryClient}>
-      <App />
+      <QueryClientProvider client={queryClient}>
+        <App />
+      </QueryClientProvider>
     </rspc.Provider>
   </React.StrictMode>
 );
