+++ create-rspc-app/templates/react_base_axum/src/utils/rspc.ts
@@ -1,0 +1,14 @@
+import { QueryClient } from "@tanstack/react-query";
+import { FetchTransport, createClient } from "@rspc/client";
+import { createReactQueryHooks } from "@rspc/react";
+
+import type { Procedures } from "./bindings"; // These are generated by rspc in Rust for you.
+
+const client = createClient<Procedures>({
+  transport: new FetchTransport("http://localhost:9000/rspc"),
+});
+
+const queryClient = new QueryClient();
+const rspc = createReactQueryHooks<Procedures>();
+
+export { rspc, client, queryClient };
