+++ create-rspc-app/templates/react_base_tauri/src/App.tsx
@@ -1,4 +1,4 @@
-import rspc from "./query.tauri";
+import { rspc } from "./utils/rspc";
 
 function App() {
   const { data } = rspc.useQuery(["version"]);
