+++ create-rspc-app/templates/react_base_tauri/src/main.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import rspc, { client, queryClient } from "./query.tauri";
+import { rspc, client, queryClient } from "./utils/rspc";
 
-ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
+ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <rspc.Provider client={client} queryClient={queryClient}>
       <App />
