+++ create-rspc-app/templates/solid_base_axum/src/App.tsx
@@ -1,16 +1,9 @@
-import type { Component } from "solid-js";
-import Comp from "./Comp";
-import rspc from "./query.axum";
+import { rspc } from "./utils/rspc";
 
-const App: Component = () => {
+function App() {
   const { data } = rspc.createQuery(() => ["version"]);
 
-  return (
-    <>
-      <h1>Hello world!!!! You are running v{data}</h1>
-      <Comp />
-    </>
-  );
-};
+  return <h1>Hello world!!!! You are running v{data}</h1>;
+}
 
 export default App;
