--- create-rspc-app/templates/solid_base_tauri/src/Comp.tsx
@@ -1,3 +1,0 @@
-export default () => {
-  return <h1>Hello world!!!</h1>;
-};
