+++ create-rspc-app/templates/solid_base_tauri/src/utils/rspc.ts
@@ -1,0 +1,14 @@
+import { QueryClient } from "@tanstack/react-query";
+import { FetchTransport, createClient } from "@rspc/client";
+import { createSolidQueryHooks } from "@rspc/solid";
+
+import type { Procedures } from "./bindings"; // These are generated by rspc in Rust for you.
+
+const client = createClient<Procedures>({
+  transport: new TauriTransport(),
+});
+
+const queryClient = new QueryClient();
+const rspc = createSolidQueryHooks<Procedures>();
+
+export { rspc, client, queryClient };
